--- src/main/scala/module1/datacollections/homework/homework1.scala
@@ -1,20 +1,0 @@
-package module1.datacollections.homework
-
-import scala.util.Random
-/*
-class BallsExperiment {
-
-  def isFirstBlackSecondWhite(): Boolean = {
-    ???
-  }
-}
-
-object BallsTest {
-  def main(args: Array[String]): Unit = {
-    val count = 10000
-    val listOfExperiments: List[BallsExperiment] = ???
-    val countOfExperiments = ???
-    val countOfPositiveExperiments: Float = countOfExperiments.count(_ == true)
-    println(countOfPositiveExperiments / count)
-  }
-}*/
